Move saveSolution from Solver.hs to Solver/Solution.hs
Makes saveSolution available in Solver/Solution.hs so it can be used
by Solver/Solve.hs or other modules without depending on the top-level
Solver module.
